backend/requirements.txt
# Framework Web
flask==3.0.0
flask-cors==4.0.0

# Banco de Dados e Autenticação
supabase==2.0.3
python-jose==3.3.0
python-multipart==0.0.6

# Processamento de Documentos
pypdf>=3.17.1
ebooklib>=0.18
beautifulsoup4>=4.12.2
html2text>=2020.1.16

# IA e Processamento de Linguagem Natural
openai>=1.3.7
tiktoken>=0.5.2
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.3

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
tqdm>=4.66.1

# Desenvolvimento e Testes
pytest>=7.4.0
black>=23.12.1
flake8>=6.1.0
mypy>=1.8.0
